Redirect up to event page if a compo is not found under it
One feedback item from 2019 mentioned finding a blank compo page.
If a request fails completely the user should see an error message and a "try again?" prompt, but an invalid compo id won't trigger that because the compos list _did_ load.
Redirecting back to the event page should help with this.